feat(schema-engine): apply review comments
