+++ schema-engine/sql-introspection-tests/tests/postgres/mod.rs
@@ -525,83 +525,3 @@ async fn commenting_stopgap(api: &mut TestApi) -> TestResult {
 
     Ok(())
 }
-
-mod check_constraints {
-    // https://www.notion.so/prismaio/Indexes-Constraints-Check-constraints-PostgreSQL-cde0bee25f6343d8bbd0f7e84932e808
-
-    use super::*;
-
-    #[test_connector(tags(Postgres), exclude(CockroachDb))]
-    async fn check_constraints_stopgap(api: &mut TestApi) -> TestResult {
-        let raw_sql = indoc! {r#"
-        CREATE TABLE products (
-            product_id serial PRIMARY KEY,
-            name text,
-            price numeric CHECK (price > 0)
-        );
-      "#};
-
-        api.raw_cmd(raw_sql).await;
-
-        let schema = expect![[r#"
-            generator client {
-              provider = "prisma-client-js"
-            }
-
-            datasource db {
-              provider = "postgresql"
-              url      = "env(TEST_DATABASE_URL)"
-            }
-
-            /// This table contains check constraints and requires additional setup for migrations. Visit https://pris.ly/d/postgres-check-constraints for more info.
-            model products {
-              product_id Int      @id @default(autoincrement())
-              name       String?
-              price      Decimal? @db.Decimal
-            }
-        "#]];
-
-        api.expect_datamodel(&schema).await;
-
-        // ensure the introspected schema is valid
-        psl::parse_schema(schema.data()).unwrap();
-
-        let expectation = expect![[r#"
-            [
-              {
-                "code": 33,
-                "message": "These constraints are not supported by the Prisma Client, because Prisma currently does not fully support check constraints. Read more: https://pris.ly/d/postgres-check-constraints",
-                "affected": [
-                  {
-                    "model": "products",
-                    "constraint": "products_price_check"
-                  }
-                ]
-              }
-            ]"#]];
-
-        api.expect_warnings(&expectation).await;
-
-        let input = indoc! { r#"
-          /// This table contains check constraints and requires additional setup for migrations. Visit https://pris.ly/d/postgres-check-constraints for more info.
-          model products {
-            product_id Int      @id @default(autoincrement())
-            name       String?
-            price      Decimal? @db.Decimal
-          }
-        "#
-        };
-
-        let expectation = expect![[r#"
-            /// This table contains check constraints and requires additional setup for migrations. Visit https://pris.ly/d/postgres-check-constraints for more info.
-            model products {
-              product_id Int      @id @default(autoincrement())
-              name       String?
-              price      Decimal? @db.Decimal
-            }
-        "#]];
-        api.expect_re_introspected_datamodel(input, expectation).await;
-
-        Ok(())
-    }
-}
